Rename dcos_api_session.py to dcos_api.py
the file doesnt solely contain the namesake class
and the module name overlaps with the class instance
name
